added full support for 16-bit audio wav sample for PS3 port

+++ engine/source/gamelib/soundmix.c
@@ -474,7 +474,7 @@ static void mixaudio(unsigned int todo)
             }
 
             // Mix a sample
-            lmusic = rmusic = sptr16[FIX_TO_INT(fp_pos)];
+            lmusic = rmusic = sptr16[FIX_TO_INT(fp_pos)]; // White Dragon: Maybe (int)(short)SwapLSB16() too?
             lmusic = (lmusic * lvolume / MAXVOLUME);
             rmusic = (rmusic * rvolume / MAXVOLUME);
             mixbuf[i++] += lmusic;
@@ -546,7 +546,7 @@ static void mixaudio(unsigned int todo)
                     #ifndef PS3
                     lmusic = rmusic = sptr16[FIX_TO_INT(fp_pos)];
                     #else
-                    lmusic = rmusic = SwapLSB16(sptr16[FIX_TO_INT(fp_pos)]);
+                    lmusic = rmusic = (int)(short)SwapLSB16(sptr16[FIX_TO_INT(fp_pos)]);
                     #endif
                     mixbuf[i++] += (lmusic * lvolume / MAXVOLUME);
                     if(vchannel[chan].channels == SOUND_MONO)
